% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pboptions.R
\name{pboptions}
\alias{pboptions}
\title{Options for \code{pbapply} functions}
\usage{
pboptions(...)
}
\arguments{
\item{...}{Pass named arguments to set \code{pbapply} options.  Run with no
arguments to list currently set options.}
}
\description{
Options for \code{pbapply} functions
}
\examples{
pboptions() # lists current arguments
pboptions(char = "=", txt.width = 80)
pboptions(style = 1) # No percent
## Reload package to reset defaults

}
\seealso{
\code{\link[utils]{txtProgressBar}}
}

